refactor: reduce aquarium maintenance risk
