+++ package.json
@@ -21,13 +21,9 @@
     "@types/three": "^0.161.2",
     "@typescript-eslint/eslint-plugin": "^8.56.1",
     "@typescript-eslint/parser": "^8.56.1",
-    "autoprefixer": "^10.4.17",
-    "daisyui": "^4.6.0",
     "eslint": "^8.56.0",
     "jsdom": "^27.4.0",
     "playwright": "^1.58.2",
-    "postcss": "^8.4.33",
-    "tailwindcss": "^3.4.1",
     "typescript": "^5.3.3",
     "vite": "^7.2.4",
     "vite-plugin-glsl": "^1.5.4",
