+++ scripts/generate-shared-aquarium-textures.mjs
@@ -1,10 +1,11 @@
 import fs from 'node:fs'
 import path from 'node:path'
 import zlib from 'node:zlib'
+import { assetPublicOutputPath } from '../src/assets/assetPathConventions.js'
 
-const plantTextureOutputDir = path.resolve('public/assets/textures/plants')
-const rockTextureOutputDir = path.resolve('public/assets/textures/rocks')
-const backdropTextureOutputDir = path.resolve('public/assets/textures/backdrop')
+const plantTextureOutputDir = path.resolve(assetPublicOutputPath('textures', 'plants'))
+const rockTextureOutputDir = path.resolve(assetPublicOutputPath('textures', 'rocks'))
+const backdropTextureOutputDir = path.resolve(assetPublicOutputPath('textures', 'backdrop'))
 
 const leafSize = 1024
 const rockSize = 1024
