+++ src/assets/assetGeneratorPaths.test.ts
@@ -1,5 +1,6 @@
 import { readFileSync } from 'node:fs'
 import { describe, expect, it } from 'vitest'
+import { assetPublicOutputPath } from './assetPathConventions.js'
 
 const readScript = (path: string): string => readFileSync(path, 'utf8')
 
@@ -7,8 +8,11 @@ describe('aquarium asset generator output paths', () => {
   it('writes generated fish assets to the same public paths used by the manifest', () => {
     const script = readScript('scripts/generate-fish-assets.mjs')
 
-    expect(script).toContain("public/assets/textures/fish")
-    expect(script).toContain("public/assets/models/fish")
+    expect(assetPublicOutputPath('textures', 'fish')).toBe('public/assets/textures/fish')
+    expect(assetPublicOutputPath('models', 'fish')).toBe('public/assets/models/fish')
+    expect(script).toContain("assetPublicOutputPath('textures', 'fish')")
+    expect(script).toContain("assetPublicOutputPath('models', 'fish')")
+    expect(script).toContain('fishAssetFileName(config.id')
     expect(script).not.toContain("public/assets/aquarium/textures/fish")
     expect(script).not.toContain("public/assets/aquarium/models/fish")
   })
@@ -22,9 +26,10 @@ describe('aquarium asset generator output paths', () => {
 
     scripts.forEach((script) => {
       expect(script).not.toContain("public/assets/aquarium/")
+      expect(script).toContain('assetPublicOutputPath')
     })
-    expect(scripts.join('\n')).toContain("public/assets/textures/plants")
-    expect(scripts.join('\n')).toContain("public/assets/models/driftwood")
-    expect(scripts.join('\n')).toContain("public/assets/models/rocks")
+    expect(assetPublicOutputPath('textures', 'plants')).toBe('public/assets/textures/plants')
+    expect(assetPublicOutputPath('models', 'driftwood')).toBe('public/assets/models/driftwood')
+    expect(assetPublicOutputPath('models', 'rocks')).toBe('public/assets/models/rocks')
   })
 })
