+++ src/assets/visualAssets.ts
@@ -1,6 +1,13 @@
 import * as THREE from 'three'
 import { GLTFLoader } from 'three/examples/jsm/loaders/GLTFLoader.js'
 import { RGBELoader } from 'three/examples/jsm/loaders/RGBELoader.js'
+import {
+  environmentAssetPath,
+  fishModelAssetPath,
+  fishTextureAssetPath,
+  modelAssetPath,
+  textureAssetPath
+} from './assetPathConventions.js'
 
 export type AssetUsageTag =
   | 'plant'
@@ -201,88 +208,88 @@ export const createAquariumAssetManifest = (
   baseUrl: string = import.meta.env.BASE_URL ?? '/'
 ): AssetManifest => ({
   textures: [
-    { id: 'leaf-diffuse', url: resolvePublicAssetUrl('assets/textures/plants/leaf-diffuse.png', baseUrl), usageTag: 'plant', lod: 'high' },
-    { id: 'leaf-alpha', url: resolvePublicAssetUrl('assets/textures/plants/leaf-alpha.png', baseUrl), usageTag: 'plant', lod: 'high', colorSpace: 'linear' },
-    { id: 'leaf-normal', url: resolvePublicAssetUrl('assets/textures/plants/leaf-normal.png', baseUrl), usageTag: 'plant', lod: 'high', colorSpace: 'linear' },
-    { id: 'leaf-roughness', url: resolvePublicAssetUrl('assets/textures/plants/leaf-roughness.png', baseUrl), usageTag: 'plant', lod: 'medium', colorSpace: 'linear' },
-    { id: 'rock-diffuse', url: resolvePublicAssetUrl('assets/textures/rocks/rock-diffuse.png', baseUrl), usageTag: 'rock', lod: 'high' },
-    { id: 'rock-normal', url: resolvePublicAssetUrl('assets/textures/rocks/rock-normal.png', baseUrl), usageTag: 'rock', lod: 'high', colorSpace: 'linear' },
-    { id: 'rock-roughness', url: resolvePublicAssetUrl('assets/textures/rocks/rock-roughness.png', baseUrl), usageTag: 'rock', lod: 'medium', colorSpace: 'linear' },
-    { id: 'driftwood-diffuse', url: resolvePublicAssetUrl('assets/textures/driftwood/driftwood-diffuse.png', baseUrl), usageTag: 'wood', lod: 'high' },
-    { id: 'driftwood-normal', url: resolvePublicAssetUrl('assets/textures/driftwood/driftwood-normal.png', baseUrl), usageTag: 'wood', lod: 'high', colorSpace: 'linear' },
-    { id: 'driftwood-roughness', url: resolvePublicAssetUrl('assets/textures/driftwood/driftwood-roughness.png', baseUrl), usageTag: 'wood', lod: 'medium', colorSpace: 'linear' },
-    { id: 'driftwood-ao', url: resolvePublicAssetUrl('assets/textures/driftwood/driftwood-ao.png', baseUrl), usageTag: 'wood', lod: 'medium', colorSpace: 'linear' },
-    { id: 'driftwood-bark-ao', url: resolvePublicAssetUrl('assets/textures/driftwood/driftwood-bark-ao.png', baseUrl), usageTag: 'wood', lod: 'high', colorSpace: 'linear' },
-    { id: 'driftwood-bark-cavity-mask', url: resolvePublicAssetUrl('assets/textures/driftwood/driftwood-bark-cavity-mask.png', baseUrl), usageTag: 'wood', lod: 'high', colorSpace: 'linear' },
-    { id: 'backdrop-depth', url: resolvePublicAssetUrl('assets/textures/backdrop/backdrop-depth.png', baseUrl), usageTag: 'backdrop', lod: 'high' },
-    { id: 'fish-tropical-basecolor', url: resolvePublicAssetUrl('assets/textures/fish/fish-tropical-basecolor.png', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-tropical-normal', url: resolvePublicAssetUrl('assets/textures/fish/fish-tropical-normal.png', baseUrl), usageTag: 'fish', lod: 'high', colorSpace: 'linear' },
-    { id: 'fish-tropical-roughness', url: resolvePublicAssetUrl('assets/textures/fish/fish-tropical-roughness.png', baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
-    { id: 'fish-tropical-alpha', url: resolvePublicAssetUrl('assets/textures/fish/fish-tropical-alpha.png', baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
-    { id: 'fish-angelfish-basecolor', url: resolvePublicAssetUrl('assets/textures/fish/fish-angelfish-basecolor.png', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-angelfish-normal', url: resolvePublicAssetUrl('assets/textures/fish/fish-angelfish-normal.png', baseUrl), usageTag: 'fish', lod: 'high', colorSpace: 'linear' },
-    { id: 'fish-angelfish-roughness', url: resolvePublicAssetUrl('assets/textures/fish/fish-angelfish-roughness.png', baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
-    { id: 'fish-angelfish-alpha', url: resolvePublicAssetUrl('assets/textures/fish/fish-angelfish-alpha.png', baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
-    { id: 'fish-butterflyfish-basecolor', url: resolvePublicAssetUrl('assets/textures/fish/fish-butterflyfish-basecolor.png', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-butterflyfish-normal', url: resolvePublicAssetUrl('assets/textures/fish/fish-butterflyfish-normal.png', baseUrl), usageTag: 'fish', lod: 'high', colorSpace: 'linear' },
-    { id: 'fish-butterflyfish-roughness', url: resolvePublicAssetUrl('assets/textures/fish/fish-butterflyfish-roughness.png', baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
-    { id: 'fish-butterflyfish-alpha', url: resolvePublicAssetUrl('assets/textures/fish/fish-butterflyfish-alpha.png', baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
-    { id: 'fish-neon-basecolor', url: resolvePublicAssetUrl('assets/textures/fish/fish-neon-basecolor.png', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-neon-normal', url: resolvePublicAssetUrl('assets/textures/fish/fish-neon-normal.png', baseUrl), usageTag: 'fish', lod: 'high', colorSpace: 'linear' },
-    { id: 'fish-neon-roughness', url: resolvePublicAssetUrl('assets/textures/fish/fish-neon-roughness.png', baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
-    { id: 'fish-neon-alpha', url: resolvePublicAssetUrl('assets/textures/fish/fish-neon-alpha.png', baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
-    { id: 'fish-goldfish-basecolor', url: resolvePublicAssetUrl('assets/textures/fish/fish-goldfish-basecolor.png', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-goldfish-normal', url: resolvePublicAssetUrl('assets/textures/fish/fish-goldfish-normal.png', baseUrl), usageTag: 'fish', lod: 'high', colorSpace: 'linear' },
-    { id: 'fish-goldfish-roughness', url: resolvePublicAssetUrl('assets/textures/fish/fish-goldfish-roughness.png', baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
-    { id: 'fish-goldfish-alpha', url: resolvePublicAssetUrl('assets/textures/fish/fish-goldfish-alpha.png', baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
-    { id: 'fish-scale-normal', url: resolvePublicAssetUrl('assets/textures/fish/fish-scale-normal.svg', baseUrl), usageTag: 'fish', lod: 'high', colorSpace: 'linear' },
-    { id: 'fish-scale-roughness', url: resolvePublicAssetUrl('assets/textures/fish/fish-scale-roughness.svg', baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
-    { id: 'substrate-sand-albedo', url: resolvePublicAssetUrl('assets/textures/substrate/substrate-sand-albedo.png', baseUrl), usageTag: 'rock', lod: 'high' },
-    { id: 'substrate-sand-normal', url: resolvePublicAssetUrl('assets/textures/substrate/substrate-sand-normal.png', baseUrl), usageTag: 'rock', lod: 'high', colorSpace: 'linear' },
-    { id: 'substrate-sand-roughness', url: resolvePublicAssetUrl('assets/textures/substrate/substrate-sand-roughness.png', baseUrl), usageTag: 'rock', lod: 'medium', colorSpace: 'linear' },
-    { id: 'substrate-sand-ao', url: resolvePublicAssetUrl('assets/textures/substrate/substrate-sand-ao.png', baseUrl), usageTag: 'rock', lod: 'medium', colorSpace: 'linear' }
+    { id: 'leaf-diffuse', url: resolvePublicAssetUrl(textureAssetPath('plants', 'leaf-diffuse.png'), baseUrl), usageTag: 'plant', lod: 'high' },
+    { id: 'leaf-alpha', url: resolvePublicAssetUrl(textureAssetPath('plants', 'leaf-alpha.png'), baseUrl), usageTag: 'plant', lod: 'high', colorSpace: 'linear' },
+    { id: 'leaf-normal', url: resolvePublicAssetUrl(textureAssetPath('plants', 'leaf-normal.png'), baseUrl), usageTag: 'plant', lod: 'high', colorSpace: 'linear' },
+    { id: 'leaf-roughness', url: resolvePublicAssetUrl(textureAssetPath('plants', 'leaf-roughness.png'), baseUrl), usageTag: 'plant', lod: 'medium', colorSpace: 'linear' },
+    { id: 'rock-diffuse', url: resolvePublicAssetUrl(textureAssetPath('rocks', 'rock-diffuse.png'), baseUrl), usageTag: 'rock', lod: 'high' },
+    { id: 'rock-normal', url: resolvePublicAssetUrl(textureAssetPath('rocks', 'rock-normal.png'), baseUrl), usageTag: 'rock', lod: 'high', colorSpace: 'linear' },
+    { id: 'rock-roughness', url: resolvePublicAssetUrl(textureAssetPath('rocks', 'rock-roughness.png'), baseUrl), usageTag: 'rock', lod: 'medium', colorSpace: 'linear' },
+    { id: 'driftwood-diffuse', url: resolvePublicAssetUrl(textureAssetPath('driftwood', 'driftwood-diffuse.png'), baseUrl), usageTag: 'wood', lod: 'high' },
+    { id: 'driftwood-normal', url: resolvePublicAssetUrl(textureAssetPath('driftwood', 'driftwood-normal.png'), baseUrl), usageTag: 'wood', lod: 'high', colorSpace: 'linear' },
+    { id: 'driftwood-roughness', url: resolvePublicAssetUrl(textureAssetPath('driftwood', 'driftwood-roughness.png'), baseUrl), usageTag: 'wood', lod: 'medium', colorSpace: 'linear' },
+    { id: 'driftwood-ao', url: resolvePublicAssetUrl(textureAssetPath('driftwood', 'driftwood-ao.png'), baseUrl), usageTag: 'wood', lod: 'medium', colorSpace: 'linear' },
+    { id: 'driftwood-bark-ao', url: resolvePublicAssetUrl(textureAssetPath('driftwood', 'driftwood-bark-ao.png'), baseUrl), usageTag: 'wood', lod: 'high', colorSpace: 'linear' },
+    { id: 'driftwood-bark-cavity-mask', url: resolvePublicAssetUrl(textureAssetPath('driftwood', 'driftwood-bark-cavity-mask.png'), baseUrl), usageTag: 'wood', lod: 'high', colorSpace: 'linear' },
+    { id: 'backdrop-depth', url: resolvePublicAssetUrl(textureAssetPath('backdrop', 'backdrop-depth.png'), baseUrl), usageTag: 'backdrop', lod: 'high' },
+    { id: 'fish-tropical-basecolor', url: resolvePublicAssetUrl(fishTextureAssetPath('tropical', 'basecolor'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-tropical-normal', url: resolvePublicAssetUrl(fishTextureAssetPath('tropical', 'normal'), baseUrl), usageTag: 'fish', lod: 'high', colorSpace: 'linear' },
+    { id: 'fish-tropical-roughness', url: resolvePublicAssetUrl(fishTextureAssetPath('tropical', 'roughness'), baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
+    { id: 'fish-tropical-alpha', url: resolvePublicAssetUrl(fishTextureAssetPath('tropical', 'alpha'), baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
+    { id: 'fish-angelfish-basecolor', url: resolvePublicAssetUrl(fishTextureAssetPath('angelfish', 'basecolor'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-angelfish-normal', url: resolvePublicAssetUrl(fishTextureAssetPath('angelfish', 'normal'), baseUrl), usageTag: 'fish', lod: 'high', colorSpace: 'linear' },
+    { id: 'fish-angelfish-roughness', url: resolvePublicAssetUrl(fishTextureAssetPath('angelfish', 'roughness'), baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
+    { id: 'fish-angelfish-alpha', url: resolvePublicAssetUrl(fishTextureAssetPath('angelfish', 'alpha'), baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
+    { id: 'fish-butterflyfish-basecolor', url: resolvePublicAssetUrl(fishTextureAssetPath('butterflyfish', 'basecolor'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-butterflyfish-normal', url: resolvePublicAssetUrl(fishTextureAssetPath('butterflyfish', 'normal'), baseUrl), usageTag: 'fish', lod: 'high', colorSpace: 'linear' },
+    { id: 'fish-butterflyfish-roughness', url: resolvePublicAssetUrl(fishTextureAssetPath('butterflyfish', 'roughness'), baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
+    { id: 'fish-butterflyfish-alpha', url: resolvePublicAssetUrl(fishTextureAssetPath('butterflyfish', 'alpha'), baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
+    { id: 'fish-neon-basecolor', url: resolvePublicAssetUrl(fishTextureAssetPath('neon', 'basecolor'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-neon-normal', url: resolvePublicAssetUrl(fishTextureAssetPath('neon', 'normal'), baseUrl), usageTag: 'fish', lod: 'high', colorSpace: 'linear' },
+    { id: 'fish-neon-roughness', url: resolvePublicAssetUrl(fishTextureAssetPath('neon', 'roughness'), baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
+    { id: 'fish-neon-alpha', url: resolvePublicAssetUrl(fishTextureAssetPath('neon', 'alpha'), baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
+    { id: 'fish-goldfish-basecolor', url: resolvePublicAssetUrl(fishTextureAssetPath('goldfish', 'basecolor'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-goldfish-normal', url: resolvePublicAssetUrl(fishTextureAssetPath('goldfish', 'normal'), baseUrl), usageTag: 'fish', lod: 'high', colorSpace: 'linear' },
+    { id: 'fish-goldfish-roughness', url: resolvePublicAssetUrl(fishTextureAssetPath('goldfish', 'roughness'), baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
+    { id: 'fish-goldfish-alpha', url: resolvePublicAssetUrl(fishTextureAssetPath('goldfish', 'alpha'), baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
+    { id: 'fish-scale-normal', url: resolvePublicAssetUrl(textureAssetPath('fish', 'fish-scale-normal.svg'), baseUrl), usageTag: 'fish', lod: 'high', colorSpace: 'linear' },
+    { id: 'fish-scale-roughness', url: resolvePublicAssetUrl(textureAssetPath('fish', 'fish-scale-roughness.svg'), baseUrl), usageTag: 'fish', lod: 'medium', colorSpace: 'linear' },
+    { id: 'substrate-sand-albedo', url: resolvePublicAssetUrl(textureAssetPath('substrate', 'substrate-sand-albedo.png'), baseUrl), usageTag: 'rock', lod: 'high' },
+    { id: 'substrate-sand-normal', url: resolvePublicAssetUrl(textureAssetPath('substrate', 'substrate-sand-normal.png'), baseUrl), usageTag: 'rock', lod: 'high', colorSpace: 'linear' },
+    { id: 'substrate-sand-roughness', url: resolvePublicAssetUrl(textureAssetPath('substrate', 'substrate-sand-roughness.png'), baseUrl), usageTag: 'rock', lod: 'medium', colorSpace: 'linear' },
+    { id: 'substrate-sand-ao', url: resolvePublicAssetUrl(textureAssetPath('substrate', 'substrate-sand-ao.png'), baseUrl), usageTag: 'rock', lod: 'medium', colorSpace: 'linear' }
   ],
   models: [
-    { id: 'fish-clownfish-school', url: resolvePublicAssetUrl('assets/models/fish/fish-clownfish-school.glb', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-clownfish-hero', url: resolvePublicAssetUrl('assets/models/fish/fish-clownfish-hero.glb', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-angelfish-school', url: resolvePublicAssetUrl('assets/models/fish/fish-angelfish-school.glb', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-angelfish-hero', url: resolvePublicAssetUrl('assets/models/fish/fish-angelfish-hero.glb', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-butterflyfish-school', url: resolvePublicAssetUrl('assets/models/fish/fish-butterflyfish-school.glb', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-butterflyfish-hero', url: resolvePublicAssetUrl('assets/models/fish/fish-butterflyfish-hero.glb', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-neon-school', url: resolvePublicAssetUrl('assets/models/fish/fish-neon-school.glb', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-neon-hero', url: resolvePublicAssetUrl('assets/models/fish/fish-neon-hero.glb', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-goldfish-school', url: resolvePublicAssetUrl('assets/models/fish/fish-goldfish-school.glb', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-goldfish-hero', url: resolvePublicAssetUrl('assets/models/fish/fish-goldfish-hero.glb', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-abeni-puffer-school', url: resolvePublicAssetUrl('assets/models/fish/fish-abeni-puffer-school.glb', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-abeni-puffer-hero', url: resolvePublicAssetUrl('assets/models/fish/fish-abeni-puffer-hero.glb', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-corydoras-school', url: resolvePublicAssetUrl('assets/models/fish/fish-corydoras-school.glb', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-corydoras-hero', url: resolvePublicAssetUrl('assets/models/fish/fish-corydoras-hero.glb', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-african-lampeye-school', url: resolvePublicAssetUrl('assets/models/fish/fish-african-lampeye-school.glb', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-african-lampeye-hero', url: resolvePublicAssetUrl('assets/models/fish/fish-african-lampeye-hero.glb', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-rasbora-heteromorpha-school', url: resolvePublicAssetUrl('assets/models/fish/fish-rasbora-heteromorpha-school.glb', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-rasbora-heteromorpha-hero', url: resolvePublicAssetUrl('assets/models/fish/fish-rasbora-heteromorpha-hero.glb', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-yamato-shrimp-school', url: resolvePublicAssetUrl('assets/models/fish/fish-yamato-shrimp-school.glb', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'fish-yamato-shrimp-hero', url: resolvePublicAssetUrl('assets/models/fish/fish-yamato-shrimp-hero.glb', baseUrl), usageTag: 'fish', lod: 'high' },
-    { id: 'plant-amazon-sword', url: resolvePublicAssetUrl('assets/models/plants/plant-amazon-sword.glb', baseUrl), usageTag: 'plant', lod: 'high' },
-    { id: 'plant-matsumo', url: resolvePublicAssetUrl('assets/models/plants/plant-matsumo.glb', baseUrl), usageTag: 'plant', lod: 'high' },
-    { id: 'plant-willow-moss', url: resolvePublicAssetUrl('assets/models/plants/plant-willow-moss.glb', baseUrl), usageTag: 'plant', lod: 'high' },
-    { id: 'plant-hygrophila-rear', url: resolvePublicAssetUrl('assets/models/plants/plant-hygrophila-rear.glb', baseUrl), usageTag: 'plant', lod: 'high' },
-    { id: 'plant-vallisneria-tall', url: resolvePublicAssetUrl('assets/models/plants/plant-vallisneria-tall.glb', baseUrl), usageTag: 'plant', lod: 'high' },
-    { id: 'driftwood-hero', url: resolvePublicAssetUrl('assets/models/driftwood/driftwood-hero.glb', baseUrl), usageTag: 'wood', lod: 'high' },
-    { id: 'driftwood-secondary-a', url: resolvePublicAssetUrl('assets/models/driftwood/driftwood-secondary-a.glb', baseUrl), usageTag: 'wood', lod: 'high' },
-    { id: 'driftwood-secondary-b', url: resolvePublicAssetUrl('assets/models/driftwood/driftwood-secondary-b.glb', baseUrl), usageTag: 'wood', lod: 'high' },
-    { id: 'driftwood-secondary-c', url: resolvePublicAssetUrl('assets/models/driftwood/driftwood-secondary-c.glb', baseUrl), usageTag: 'wood', lod: 'high' },
-    { id: 'rock-ridge-hero', url: resolvePublicAssetUrl('assets/models/rocks/rock-ridge-hero.glb', baseUrl), usageTag: 'rock', lod: 'high' },
-    { id: 'rock-support-a', url: resolvePublicAssetUrl('assets/models/rocks/rock-support-a.glb', baseUrl), usageTag: 'rock', lod: 'high' },
-    { id: 'rock-support-b', url: resolvePublicAssetUrl('assets/models/rocks/rock-support-b.glb', baseUrl), usageTag: 'rock', lod: 'high' },
-    { id: 'rock-support-c', url: resolvePublicAssetUrl('assets/models/rocks/rock-support-c.glb', baseUrl), usageTag: 'rock', lod: 'high' },
-    { id: 'rock-pebble-cluster', url: resolvePublicAssetUrl('assets/models/rocks/rock-pebble-cluster.glb', baseUrl), usageTag: 'rock', lod: 'high' },
-    { id: 'rock-lava-base-cluster-a', url: resolvePublicAssetUrl('assets/models/rocks/rock-lava-base-cluster-a.glb', baseUrl), usageTag: 'rock', lod: 'high' },
-    { id: 'rock-lava-base-cluster-b', url: resolvePublicAssetUrl('assets/models/rocks/rock-lava-base-cluster-b.glb', baseUrl), usageTag: 'rock', lod: 'high' },
-    { id: 'rock-lava-transition-chips', url: resolvePublicAssetUrl('assets/models/rocks/rock-lava-transition-chips.glb', baseUrl), usageTag: 'rock', lod: 'high' }
+    { id: 'fish-clownfish-school', url: resolvePublicAssetUrl(fishModelAssetPath('clownfish', 'school'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-clownfish-hero', url: resolvePublicAssetUrl(fishModelAssetPath('clownfish', 'hero'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-angelfish-school', url: resolvePublicAssetUrl(fishModelAssetPath('angelfish', 'school'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-angelfish-hero', url: resolvePublicAssetUrl(fishModelAssetPath('angelfish', 'hero'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-butterflyfish-school', url: resolvePublicAssetUrl(fishModelAssetPath('butterflyfish', 'school'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-butterflyfish-hero', url: resolvePublicAssetUrl(fishModelAssetPath('butterflyfish', 'hero'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-neon-school', url: resolvePublicAssetUrl(fishModelAssetPath('neon', 'school'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-neon-hero', url: resolvePublicAssetUrl(fishModelAssetPath('neon', 'hero'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-goldfish-school', url: resolvePublicAssetUrl(fishModelAssetPath('goldfish', 'school'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-goldfish-hero', url: resolvePublicAssetUrl(fishModelAssetPath('goldfish', 'hero'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-abeni-puffer-school', url: resolvePublicAssetUrl(fishModelAssetPath('abeni-puffer', 'school'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-abeni-puffer-hero', url: resolvePublicAssetUrl(fishModelAssetPath('abeni-puffer', 'hero'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-corydoras-school', url: resolvePublicAssetUrl(fishModelAssetPath('corydoras', 'school'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-corydoras-hero', url: resolvePublicAssetUrl(fishModelAssetPath('corydoras', 'hero'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-african-lampeye-school', url: resolvePublicAssetUrl(fishModelAssetPath('african-lampeye', 'school'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-african-lampeye-hero', url: resolvePublicAssetUrl(fishModelAssetPath('african-lampeye', 'hero'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-rasbora-heteromorpha-school', url: resolvePublicAssetUrl(fishModelAssetPath('rasbora-heteromorpha', 'school'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-rasbora-heteromorpha-hero', url: resolvePublicAssetUrl(fishModelAssetPath('rasbora-heteromorpha', 'hero'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-yamato-shrimp-school', url: resolvePublicAssetUrl(fishModelAssetPath('yamato-shrimp', 'school'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'fish-yamato-shrimp-hero', url: resolvePublicAssetUrl(fishModelAssetPath('yamato-shrimp', 'hero'), baseUrl), usageTag: 'fish', lod: 'high' },
+    { id: 'plant-amazon-sword', url: resolvePublicAssetUrl(modelAssetPath('plants', 'plant-amazon-sword.glb'), baseUrl), usageTag: 'plant', lod: 'high' },
+    { id: 'plant-matsumo', url: resolvePublicAssetUrl(modelAssetPath('plants', 'plant-matsumo.glb'), baseUrl), usageTag: 'plant', lod: 'high' },
+    { id: 'plant-willow-moss', url: resolvePublicAssetUrl(modelAssetPath('plants', 'plant-willow-moss.glb'), baseUrl), usageTag: 'plant', lod: 'high' },
+    { id: 'plant-hygrophila-rear', url: resolvePublicAssetUrl(modelAssetPath('plants', 'plant-hygrophila-rear.glb'), baseUrl), usageTag: 'plant', lod: 'high' },
+    { id: 'plant-vallisneria-tall', url: resolvePublicAssetUrl(modelAssetPath('plants', 'plant-vallisneria-tall.glb'), baseUrl), usageTag: 'plant', lod: 'high' },
+    { id: 'driftwood-hero', url: resolvePublicAssetUrl(modelAssetPath('driftwood', 'driftwood-hero.glb'), baseUrl), usageTag: 'wood', lod: 'high' },
+    { id: 'driftwood-secondary-a', url: resolvePublicAssetUrl(modelAssetPath('driftwood', 'driftwood-secondary-a.glb'), baseUrl), usageTag: 'wood', lod: 'high' },
+    { id: 'driftwood-secondary-b', url: resolvePublicAssetUrl(modelAssetPath('driftwood', 'driftwood-secondary-b.glb'), baseUrl), usageTag: 'wood', lod: 'high' },
+    { id: 'driftwood-secondary-c', url: resolvePublicAssetUrl(modelAssetPath('driftwood', 'driftwood-secondary-c.glb'), baseUrl), usageTag: 'wood', lod: 'high' },
+    { id: 'rock-ridge-hero', url: resolvePublicAssetUrl(modelAssetPath('rocks', 'rock-ridge-hero.glb'), baseUrl), usageTag: 'rock', lod: 'high' },
+    { id: 'rock-support-a', url: resolvePublicAssetUrl(modelAssetPath('rocks', 'rock-support-a.glb'), baseUrl), usageTag: 'rock', lod: 'high' },
+    { id: 'rock-support-b', url: resolvePublicAssetUrl(modelAssetPath('rocks', 'rock-support-b.glb'), baseUrl), usageTag: 'rock', lod: 'high' },
+    { id: 'rock-support-c', url: resolvePublicAssetUrl(modelAssetPath('rocks', 'rock-support-c.glb'), baseUrl), usageTag: 'rock', lod: 'high' },
+    { id: 'rock-pebble-cluster', url: resolvePublicAssetUrl(modelAssetPath('rocks', 'rock-pebble-cluster.glb'), baseUrl), usageTag: 'rock', lod: 'high' },
+    { id: 'rock-lava-base-cluster-a', url: resolvePublicAssetUrl(modelAssetPath('rocks', 'rock-lava-base-cluster-a.glb'), baseUrl), usageTag: 'rock', lod: 'high' },
+    { id: 'rock-lava-base-cluster-b', url: resolvePublicAssetUrl(modelAssetPath('rocks', 'rock-lava-base-cluster-b.glb'), baseUrl), usageTag: 'rock', lod: 'high' },
+    { id: 'rock-lava-transition-chips', url: resolvePublicAssetUrl(modelAssetPath('rocks', 'rock-lava-transition-chips.glb'), baseUrl), usageTag: 'rock', lod: 'high' }
   ],
   environment: [
-    { id: 'aquarium-hdri', url: resolvePublicAssetUrl('assets/environment/aquarium-hdri.hdr', baseUrl), usageTag: 'environment', lod: 'high' }
+    { id: 'aquarium-hdri', url: resolvePublicAssetUrl(environmentAssetPath('aquarium-hdri.hdr'), baseUrl), usageTag: 'environment', lod: 'high' }
   ]
 })
