+++ src/components/substrateGeometry.ts
@@ -1,0 +1,9 @@
+export const SUBSTRATE_GEOMETRY_SEGMENTS = {
+  topWidth: 160,
+  topDepth: 112
+} as const
+
+export const SUBSTRATE_VISUAL_FOOTPRINT_SCALE = {
+  width: 3.62,
+  depth: 3.68
+} as const
